TRIPLEX-830 Добавлены PlaygroundExample для Calendar, Cards, Checkbox, CheckboxTree, CheckboxTreeExtended
