+++ stories/Calendar/examples/index.ts
@@ -2,3 +2,6 @@ export * from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export * from "./PickTypesExample";
 export { default as PickTypesExampleSource } from "./PickTypesExample?raw";
+
+export * from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
