+++ stories/Cards/CardAction.stories.tsx
@@ -1,25 +1,12 @@
-import React, { useState } from "react";
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { action } from "storybook/actions";
 import {
     CardAction,
     ICardActionProps,
-    EButtonTheme,
     ECardTheme,
     ECardRoundingSize,
     ECardContentPaddingSize,
-    Title,
-    ETitleSize,
-    EFontWeightTitle,
-    Button,
-    EComponentSize,
-    Text,
-    ETextSize,
-    EFontType,
-    Link,
-    Gap,
 } from "@sberbusiness/triplex-next";
-import { DefaulticonStrokePrdIcon20 } from "@sberbusiness/icons-next";
 import {
     Title as DocsTitle,
     Description,
@@ -30,18 +17,16 @@ import {
     ArgTypes,
 } from "@storybook/addon-docs/blocks";
 import {
-    CardActionDefaultExample,
-    CardActionDefaultExampleSource,
-    CardActionPaddingSizesExample,
-    CardActionPaddingSizesExampleSource,
-    CardActionRoundingSizesExample,
-    CardActionRoundingSizesExampleSource,
-    CardActionThemesExample,
-    CardActionThemesExampleSource,
+    DefaultExample,
+    DefaultExampleSource,
+    PaddingSizesExample,
+    PaddingSizesExampleSource,
+    PlaygroundExample,
+    RoundingSizesExample,
+    RoundingSizesExampleSource,
+    ThemesExample,
+    ThemesExampleSource,
 } from "./examples/CardAction";
-
-/** Высота блока Media */
-const MEDIA_HEIGHT = "129px";
 
 type TCardActionPlaygroundProps = Pick<
     ICardActionProps,
@@ -122,74 +107,7 @@ export const Playground: StoryObj<TCardActionPlaygroundProps> = {
         theme: ECardTheme.GENERAL,
         selected: false,
     },
-    render: (args) => {
-        const { paddingSize, theme, ...cardArgs } = args;
-
-        const [isSelected, setIsSelected] = useState(args?.selected ?? false);
-
-        const handleToggle = (selected: boolean) => {
-            setIsSelected(selected);
-            args.onToggle?.(selected);
-            args.toggle?.(selected);
-            action("onToggle")(selected);
-            action("toggle")(selected);
-        };
-
-        const buttomTheme = isSelected ? EButtonTheme.SECONDARY_LIGHT : EButtonTheme.SECONDARY;
-        const isGeneralTheme = theme === ECardTheme.GENERAL;
-
-        return (
-            <div style={{ width: "216px" }}>
-                <CardAction {...cardArgs} theme={theme} selected={isSelected} toggle={handleToggle}>
-                    <CardAction.Media
-                        style={{ backgroundImage: "url(assets/images/evotor.png)", height: MEDIA_HEIGHT }}
-                    />
-                    <CardAction.Content paddingSize={paddingSize}>
-                        <CardAction.Content.Header>
-                            <Title tag="div" size={ETitleSize.H3} weight={EFontWeightTitle.REGULAR}>
-                                Title text
-                            </Title>
-                        </CardAction.Content.Header>
-                        <CardAction.Content.Body>
-                            <div style={{ display: "flex", alignItems: "center" }}>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                                <Text size={ETextSize.B3} style={{ marginLeft: "8px" }}>
-                                    List item text
-                                </Text>
-                            </div>
-                            <div style={{ display: "flex", alignItems: "center", marginTop: "16px" }}>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                                <Text size={ETextSize.B3} style={{ marginLeft: "8px" }}>
-                                    List item text
-                                </Text>
-                            </div>
-                            <div style={{ display: "flex", alignItems: "center", marginTop: "16px" }}>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                                <Text size={ETextSize.B3} style={{ marginLeft: "8px" }}>
-                                    List item text
-                                </Text>
-                            </div>
-                            <Gap size={8} />
-                            <Text tag="div" type={EFontType.SECONDARY} size={ETextSize.B4}>
-                                This message provides additional context or highlights important information to note.
-                            </Text>
-                            <Gap size={8} />
-                            <Text tag="div" type={EFontType.PRIMARY} size={ETextSize.B3}>
-                                <Link onClick={() => {}}>Link text</Link>
-                            </Text>
-                        </CardAction.Content.Body>
-                        {isGeneralTheme && (
-                            <CardAction.Content.Footer>
-                                <Button theme={buttomTheme} size={EComponentSize.SM}>
-                                    Button text
-                                </Button>
-                            </CardAction.Content.Footer>
-                        )}
-                    </CardAction.Content>
-                </CardAction>
-            </div>
-        );
-    },
+    render: PlaygroundExample,
 };
 
 export const Default: StoryObj<TCardActionPlaygroundProps> = {
@@ -197,12 +115,12 @@ export const Default: StoryObj<TCardActionPlaygroundProps> = {
         controls: { disable: true },
         docs: {
             source: {
-                code: CardActionDefaultExampleSource,
+                code: DefaultExampleSource,
                 language: "tsx",
             },
         },
     },
-    render: CardActionDefaultExample,
+    render: DefaultExample,
 };
 
 export const Themes: StoryObj<TCardActionPlaygroundProps> = {
@@ -210,12 +128,12 @@ export const Themes: StoryObj<TCardActionPlaygroundProps> = {
         controls: { disable: true },
         docs: {
             source: {
-                code: CardActionThemesExampleSource,
+                code: ThemesExampleSource,
                 language: "tsx",
             },
         },
     },
-    render: CardActionThemesExample,
+    render: ThemesExample,
 };
 
 export const PaddingSizes: StoryObj<TCardActionPlaygroundProps> = {
@@ -223,12 +141,12 @@ export const PaddingSizes: StoryObj<TCardActionPlaygroundProps> = {
         controls: { disable: true },
         docs: {
             source: {
-                code: CardActionPaddingSizesExampleSource,
+                code: PaddingSizesExampleSource,
                 language: "tsx",
             },
         },
     },
-    render: CardActionPaddingSizesExample,
+    render: PaddingSizesExample,
 };
 
 export const RoundingSizes: StoryObj<TCardActionPlaygroundProps> = {
@@ -236,10 +154,10 @@ export const RoundingSizes: StoryObj<TCardActionPlaygroundProps> = {
         controls: { disable: true },
         docs: {
             source: {
-                code: CardActionRoundingSizesExampleSource,
+                code: RoundingSizesExampleSource,
                 language: "tsx",
             },
         },
     },
-    render: CardActionRoundingSizesExample,
+    render: RoundingSizesExample,
 };
