+++ stories/Cards/examples/CardAction/index.ts
@@ -1,8 +1,10 @@
-export * from "./CardActionDefaultExample";
-export { default as CardActionDefaultExampleSource } from "./CardActionDefaultExample?raw";
-export * from "./CardActionThemesExample";
-export { default as CardActionThemesExampleSource } from "./CardActionThemesExample?raw";
-export * from "./CardActionPaddingSizesExample";
-export { default as CardActionPaddingSizesExampleSource } from "./CardActionPaddingSizesExample?raw";
-export * from "./CardActionRoundingSizesExample";
-export { default as CardActionRoundingSizesExampleSource } from "./CardActionRoundingSizesExample?raw";
+export * from "./DefaultExample";
+export { default as DefaultExampleSource } from "./DefaultExample?raw";
+export * from "./ThemesExample";
+export { default as ThemesExampleSource } from "./ThemesExample?raw";
+export * from "./PaddingSizesExample";
+export { default as PaddingSizesExampleSource } from "./PaddingSizesExample?raw";
+export * from "./RoundingSizesExample";
+export { default as RoundingSizesExampleSource } from "./RoundingSizesExample?raw";
+export * from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
