+++ stories/Cards/examples/CardStatic/index.ts
@@ -1,8 +1,10 @@
-export * from "./CardStaticDefaultExample";
-export { default as CardStaticDefaultExampleSource } from "./CardStaticDefaultExample?raw";
-export * from "./CardStaticThemesExample";
-export { default as CardStaticThemesExampleSource } from "./CardStaticThemesExample?raw";
-export * from "./CardStaticPaddingSizesExample";
-export { default as CardStaticPaddingSizesExampleSource } from "./CardStaticPaddingSizesExample?raw";
-export * from "./CardStaticRoundingSizesExample";
-export { default as CardStaticRoundingSizesExampleSource } from "./CardStaticRoundingSizesExample?raw";
+export * from "./DefaultExample";
+export { default as DefaultExampleSource } from "./DefaultExample?raw";
+export * from "./ThemesExample";
+export { default as ThemesExampleSource } from "./ThemesExample?raw";
+export * from "./PaddingSizesExample";
+export { default as PaddingSizesExampleSource } from "./PaddingSizesExample?raw";
+export * from "./RoundingSizesExample";
+export { default as RoundingSizesExampleSource } from "./RoundingSizesExample?raw";
+export * from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
