+++ stories/Checkbox/examples/PlaygroundExample.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import { Checkbox } from "@sberbusiness/triplex-next";
+
+export const PlaygroundExample = (args: React.ComponentProps<typeof Checkbox>) => <Checkbox {...args} />;
