+++ stories/CheckboxTreeExtended/examples/index.ts
@@ -2,3 +2,5 @@ export * from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export * from "./SizesExample";
 export { default as SizesExampleSource } from "./SizesExample?raw";
+export * from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
