+++ stories/Chips/ChipMultiselect.stories.tsx
@@ -140,7 +140,6 @@ export const Playground: StoryObj<typeof ChipMultiselect> = {
 
                 return CHIP_MULTISELECT_OPTIONS.filter((opt) => opt.label.toLowerCase().includes(lower));
             }, [filter]);
-
             const handleToggle = (optionId: string, checked: boolean) => {
                 setSelectedIds((prev) => {
                     if (checked) {
@@ -170,7 +169,6 @@ export const Playground: StoryObj<typeof ChipMultiselect> = {
                     ))}
                 </CheckboxYGroup>
             );
-
             const renderDropdownContent = () => {
                 const renderCheckboxes = !filter.trim() || visibleOptions.length > 0;
                 return renderCheckboxes ? (
@@ -198,7 +196,6 @@ export const Playground: StoryObj<typeof ChipMultiselect> = {
             const handleFilterChange = (event: React.ChangeEvent<HTMLInputElement>) => {
                 setFilter(event.target.value);
             };
-
             const renderDropdown = ({ opened, setOpened, targetRef, dropdownRef }) => (
                 <MultiselectField.Dropdown
                     opened={opened}
