+++ stories/Chips/examples/Chip/DefaultExample.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Chip } from "@sberbusiness/triplex-next";
 
-export const ChipDefaultExample = () => {
+export const DefaultExample = () => {
     const [selected, setSelected] = useState(false);
     return (
         <Chip selected={selected} onClick={() => setSelected((s) => !s)}>
