+++ stories/Chips/examples/Chip/StatesExample.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Chip } from "@sberbusiness/triplex-next";
 
-export const ChipStatesExample = () => (
+export const StatesExample = () => (
     <div style={{ display: "flex", gap: 12 }}>
         <Chip selected>Selected</Chip>
         <Chip disabled>Disabled</Chip>
