+++ stories/Chips/examples/ChipGroup/index.ts
@@ -1,6 +1,8 @@
-export { ChipGroupDefaultExample as DefaultExample } from "./ChipGroupDefaultExample";
-export { default as DefaultExampleSource } from "./ChipGroupDefaultExample?raw";
-export { ChipGroupSizesExample as SizesExample } from "./ChipGroupSizesExample";
-export { default as SizesExampleSource } from "./ChipGroupSizesExample?raw";
-export { ChipGroupOneLineExample as OneLineExample } from "./ChipGroupOneLineExample";
-export { default as OneLineExampleSource } from "./ChipGroupOneLineExample?raw";
+export { DefaultExample } from "./DefaultExample";
+export { default as DefaultExampleSource } from "./DefaultExample?raw";
+export { SizesExample } from "./SizesExample";
+export { default as SizesExampleSource } from "./SizesExample?raw";
+export { OneLineExample } from "./OneLineExample";
+export { default as OneLineExampleSource } from "./OneLineExample?raw";
+export { PlaygroundExample } from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
