+++ stories/Chips/examples/ChipIcon/PlaygroundExample.tsx
@@ -1,0 +1,13 @@
+import React, { useState } from "react";
+import { DefaulticonStrokePrdIcon24 } from "@sberbusiness/icons-next";
+import { ChipIcon } from "@sberbusiness/triplex-next";
+
+export const PlaygroundExample = (args: React.ComponentProps<typeof ChipIcon>) => {
+    const [selected, setSelected] = useState(false);
+
+    return (
+        <ChipIcon {...args} selected={selected} onClick={() => setSelected((s) => !s)}>
+            <DefaulticonStrokePrdIcon24 paletteIndex={selected ? 6 : 5} />
+        </ChipIcon>
+    );
+};
